requirements.txt
-f /usr/share/pip-wheels
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-adminlte3==0.1.6
django-import-export==2.4.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.25.11
xlrd==2.0.1
xlwt==1.3.0
